+++ classes/Middleware/PsrMiddleware.php
@@ -3,9 +3,9 @@
 namespace Neat\Http\Server\Middleware;
 
 use Neat\Http\Response;
-use Neat\Http\ServerRequest;
 use Neat\Http\Server\Handler;
 use Neat\Http\Server\Middleware;
+use Neat\Http\Server\Request;
 use Psr\Http\Server\MiddlewareInterface;
 
 class PsrMiddleware implements Middleware
@@ -32,11 +32,11 @@ class PsrMiddleware implements Middleware
     }
 
     /**
-     * @param ServerRequest $request
-     * @param Handler       $handler
+     * @param Request $request
+     * @param Handler $handler
      * @return Response
      */
-    public function process(ServerRequest $request, Handler $handler): Response
+    public function process(Request $request, Handler $handler): Response
     {
         $psrHandler
             = $handler instanceof Handler\PsrHandler