+++ classes/Middleware/PsrWrapper.php
@@ -2,9 +2,9 @@
 
 namespace Neat\Http\Server\Middleware;
 
-use Neat\Http\ServerRequest;
 use Neat\Http\Server\Handler\PsrHandler;
 use Neat\Http\Server\Middleware;
+use Neat\Http\Server\Request;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
@@ -30,6 +30,6 @@ class PsrWrapper implements MiddlewareInterface
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
-        return $this->middleware->process(new ServerRequest($request), new PsrHandler($handler))->psr();
+        return $this->middleware->process(new Request($request), new PsrHandler($handler))->psr();
     }
 }