+++ tests/Middleware/CallableMiddlewareTest.php
@@ -3,9 +3,9 @@
 namespace Neat\Http\Server\Test\Middleware;
 
 use Neat\Http\Response;
-use Neat\Http\ServerRequest;
 use Neat\Http\Server\Handler;
 use Neat\Http\Server\Middleware\CallableMiddleware;
+use Neat\Http\Server\Request;
 use Neat\Http\Server\Test\CallableMock;
 use PHPUnit\Framework\TestCase;
 
@@ -23,7 +23,7 @@ class CallableMiddlewareTest extends TestCase
 
     public function testProcess()
     {
-        $request  = $this->createMock(ServerRequest::class);
+        $request  = $this->createMock(Request::class);
         $response = $this->createMock(Response::class);
         $handler  = $this->createMock(Handler::class);